Combine height and weight questions
