+++ client/containers/summary-handler.jsx
@@ -15,8 +15,7 @@ import {
   Sex,
   EyeColor,
   HairColor,
-  Height,
-  Weight,
+  TraitsHeightWeight,
   OrganDonation,
   SocialSecurity,
   PrivilegeRemovedHistory,
@@ -44,8 +43,7 @@ const SummaryHandler = (props) => {
     <Sex sex={props.sex} key='sex' />,
     <EyeColor eyeColor={props.eyeColor} key='eye-color' />,
     <HairColor hairColor={props.hairColor} key='hair-color' />,
-    <Height height={props.height} key='height' />,
-    <Weight weight={props.weight} key='weight' />,
+    <TraitsHeightWeight traitsHeightWeight={props.traitsHeightWeight} key='traits-height-weight' />,
     <OrganDonation organDonation={props.organDonation} key='organ-donation' />,
     <SocialSecurity socialSecurity={props.socialSecurity} key='social-security' />,
     <PrivilegeRemovedHistory privilegeRemovedHistory={props.privilegeRemovedHistory} key='privilege-removed-history' />,
