+++ client/presentations/home.jsx
@@ -23,8 +23,7 @@ const linkData = [
   {description: 'sex',                                    path: '/about-me/sex'},
   {description: 'eye color',                              path: '/about-me/appearance/eye'},
   {description: 'hair color',                             path: '/about-me/appearance/hair'},
-  {description: 'height',                                 path: '/about-me/height'},
-  {description: 'weight',                                 path: '/about-me/weight'},
+  {description: 'traits height and weight',               path: '/about-me/traits-height-weight'},
   {description: 'social security',                        path: '/about-me/social-security'},
   {description: 'organ donation',                         path: '/about-me/organ-donation'},
   {description: 'has existing DL/ID',                     path: '/about-me/license-history'},
