+++ client/reducers/application/update-traits-height-weight.js
@@ -6,8 +6,9 @@ import formObjectReducer from './form-object-reducer';
 function defaultState() {
   return {
     feet: '',
-    inches: ''
+    inches: '',
+    weight: ''
   };
 }
 
-export default formObjectReducer(defaultState, TYPES.UPDATE_HEIGHT);
+export default formObjectReducer(defaultState, TYPES.UPDATE_TRAITS_HEIGHT_WEIGHT);
