--- client/reducers/application/update-weight.js
@@ -1,10 +1,0 @@
-'use strict';
-
-import { TYPES } from '../../actions';
-import formValueReducer from './form-value-reducer';
-
-function defaultState() {
-  return '';
-}
-
-export default formValueReducer(defaultState, TYPES.UPDATE_WEIGHT);
