+++ test/features/enter-hair-color.feature
@@ -12,7 +12,7 @@ Feature: I want to select my hair color
     Then I will see the hair color I chose is selected
     Then I will see that the Continue button is no longer disabled
     When I click to submit
-    Then I will be on the page for entering my height
+    Then I will be on the page for entering my height and weight
     And I go to the page with my summary
     Then I will see my hair color in the summary
