+++ test/features/enter-traits-height-weight.feature
@@ -5,22 +5,23 @@ Feature:  I want to enter my height
 
   Scenario: Entering my weight and saving
     Given I go to the new online DL application page
-    When I visit the height page
-    Then I will see a field for height feet and inches
+    When I visit the traits height and weight page
+    Then I will see a field for height and weight
     And I will see that the Continue button is disabled
     When I enter my feet
-    Then I will see that the Continue button is no longer disabled
     When I enter my inches
+    When I enter my weight
+    Then I will see that the Continue button is no longer disabled
     And I click to submit
-    Then I will be on the page for entering my weight
+    Then I will be on the page for entering my social security
     And I go to the page with my summary
     Then I will see my height on that summary
 
   Scenario: Updating weight data
     Given I go to the new online DL application page
-    And I have already entered my height into the form
-    When I visit the height page
-    Then I will see the height I entered
+    And I have already entered my height and weight into the form
+    When I visit the traits height and weight page
+    Then I will see the height and weight I entered
     When I change my inches
     And I click to submit
     And I go to the page with my summary
